move code to ./src
