+++ src/emitters/bad_gateway_waps.json
