+++ src/emitters/cell_partners.py
