+++ src/utilities/config/dashboards.kibana
