+++ src/utilities/config/index_patterns.kibana
