+++ src/utilities/config/intro_workpad.kibana
