+++ src/utilities/config/network_events_mapping.json
