+++ src/utilities/config/spaces_privs.kibana
