+++ src/utilities/config/visualizations.kibana
